vmware: Support guestinfo.ignition.config.* variables
The Ignition data and the cloud-config data resides at the same place.
There was a name change from coreos.config to ignition.config in upstream
Ignition. This means that if only ignition.config is used as guestinfo
variable only Ignition configs are currently accepted while both should
be supported.
Support the ignition.config variables and fall back to coreos.config
variables the same way Ignition does, so that the user or a provisioning
tool does not need to care which config format is used in the guestinfo
variable.