user-configdrive: Restore ordering with configdrive cloudinit
Since we don't enable the oem cloudinit service from the initrd but from
the real system, add an additional ordering for the service that enables
the oem cloudinit service.
We could also move the cloudinit enablement back to the initrd to avoid
this race but maybe this here is enough.