units: Don't use configdrive on DigitalOcean
For DigitalOcean we have the oem-cloudinit service and if the
configdrive gets used in addition, it will process the metadata
unconditionally and set the hostname even if afterburn already prepared
the hostname and Ignition maybe overwrote it with a static value.
Don't use the configdrive trigger at all on DigitalOcean because both
the metadata is covered by afterburn and if cloudinit userdata is found,
oem-cloudinit will run.