units: Workaround for user-configvirtfs.service
As Jeremi noticed with the current Flatcar Stable 3510.2.7 and a custom
Docker brought in as systemd-sysext image, the qemu VM didn't pick up
the SSH keys anymore. This couldn't be reproduced with Alpha, though.
The culprit is not completely known but it seems to be a race condition
that makes the Wants= dependency of media-configvirtfs.mount not take
action.
To work around the problem, use Upholds= which eventually starts
user-configvirtfs.service.