Restore old behavior of not setting empty hostnames
The refactoring in https://github.com/flatcar/coreos-cloudinit/pull/21
caused hostnames to be set unconditionally compared to the old behavior
of only setting the hostname if it not empty.

When running coreos-cloudinit with datasources that do not provide
metadata such as the `file` datasource, the refactored code caused the
hostname to always be reset to `localhost`. This leads to various
problems like preventing k8s nodes from joining their cluster.

This change restores the old behavior by not applying empty hostnames.

Fixes https://github.com/flatcar/Flatcar/issues/1262